pyproject.toml
[project]
name = "mlx-server"
version = "0.1.0"
description = "OpenAI-compatible API server for Gemma 3 4B via MLX"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "mlx>=0.22.0",
    "mlx-lm>=0.22.0",
    "mlx-vlm>=0.1.18",
    "pydantic>=2.0.0",
    "sse-starlette>=2.0.0",
    "pillow>=10.0.0",
    "httpx>=0.27.0",
    "torchvision>=0.20.0",
]

[project.scripts]
mlx-server = "mlx_server.main:main"
